pyproject.toml
[tool.poetry]
name = "mobile-communications"
version = "0.1.0"
description = "Site that allows you to call someone"
authors = ["Denis Romodanov <darvenommm@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.2"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
coverage = "^7.5.0"
mypy = "^1.10.0"
flake8 = "^7.0.0"
wemake-python-styleguide = "^0.19.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py312"]
line-length = 100
